reuse common functional test code between app ext and app tests
