+++ .circleci/config.yml
@@ -35,7 +35,7 @@ jobs:
       # install CocoaPods - using default CocoaPods version, not the bundle
       - run:
           name: Repo Update & Install CocoaPods
-          command: bundle exec pod install
+          command: bundle exec pod install --repo-update
 
       # save pods related files
       - save_cache:
