+++ AnalyticsSampleAppExt/FileProviderExtension.swift
@@ -18,10 +18,10 @@ import AEPIdentity
 import AEPAnalytics
 
 class FileProviderExtension: NSFileProviderExtension {
-    
+
     var fileManager = FileManager()
     private let LAUNCH_ENVIRONMENT_FILE_ID = ""
-    
+
     override init() {
         super.init()
         MobileCore.setLogLevel(.trace)
@@ -35,38 +35,38 @@ class FileProviderExtension: NSFileProviderExtension {
         })
 
     }
-    
+
     override func item(for identifier: NSFileProviderItemIdentifier) throws -> NSFileProviderItem {
         // resolve the given identifier to a record in the model
-        
+
         // TODO: implement the actual lookup
         return FileProviderItem()
     }
-    
+
     override func urlForItem(withPersistentIdentifier identifier: NSFileProviderItemIdentifier) -> URL? {
         // resolve the given identifier to a file on disk
         guard let item = try? item(for: identifier) else {
             return nil
         }
-        
+
         // in this implementation, all paths are structured as <base storage directory>/<item identifier>/<item file name>
         let manager = NSFileProviderManager.default
         let perItemDirectory = manager.documentStorageURL.appendingPathComponent(identifier.rawValue, isDirectory: true)
-        
-        return perItemDirectory.appendingPathComponent(item.filename, isDirectory:false)
+
+        return perItemDirectory.appendingPathComponent(item.filename, isDirectory: false)
     }
-    
+
     override func persistentIdentifierForItem(at url: URL) -> NSFileProviderItemIdentifier? {
         // resolve the given URL to a persistent identifier using a database
         let pathComponents = url.pathComponents
-        
+
         // exploit the fact that the path structure has been defined as
         // <base storage directory>/<item identifier>/<item file name> above
         assert(pathComponents.count > 2)
-        
+
         return NSFileProviderItemIdentifier(pathComponents[pathComponents.count - 2])
     }
-    
+
     override func providePlaceholder(at url: URL, completionHandler: @escaping (Error?) -> Void) {
         guard let identifier = persistentIdentifierForItem(at: url) else {
             completionHandler(NSFileProviderError(.noSuchItem))
@@ -76,7 +76,7 @@ class FileProviderExtension: NSFileProviderExtension {
         do {
             let fileProviderItem = try item(for: identifier)
             let placeholderURL = NSFileProviderManager.placeholderURL(for: url)
-            try NSFileProviderManager.writePlaceholder(at: placeholderURL,withMetadata: fileProviderItem)
+            try NSFileProviderManager.writePlaceholder(at: placeholderURL, withMetadata: fileProviderItem)
             completionHandler(nil)
         } catch let error {
             completionHandler(error)
@@ -85,7 +85,6 @@ class FileProviderExtension: NSFileProviderExtension {
 
     override func startProvidingItem(at url: URL, completionHandler: @escaping ((_ error: Error?) -> Void)) {
         // Should ensure that the actual file is in the position returned by URLForItemWithIdentifier:, then call the completion handler
-        
         /* TODO:
          This is one of the main entry points of the file provider. We need to check whether the file already exists on disk,
          whether we know of a more recent version of the file, and implement a policy for these cases. Pseudocode:
@@ -109,14 +108,12 @@ class FileProviderExtension: NSFileProviderExtension {
              }
          }
          */
-        
-        completionHandler(NSError(domain: NSCocoaErrorDomain, code: NSFeatureUnsupportedError, userInfo:[:]))
+        completionHandler(NSError(domain: NSCocoaErrorDomain, code: NSFeatureUnsupportedError, userInfo: [:]))
     }
-    
-    
+
     override func itemChanged(at url: URL) {
         // Called at some point after the file has changed; the provider may then trigger an upload
-        
+
         /* TODO:
          - mark file at <url> as needing an update in the model
          - if there are existing NSURLSessionTasks uploading this file, cancel them
@@ -124,16 +121,16 @@ class FileProviderExtension: NSFileProviderExtension {
          - register the NSURLSessionTask with NSFileProviderManager to provide progress updates
          */
     }
-    
+
     override func stopProvidingItem(at url: URL) {
         // Called after the last claim to the file has been released. At this point, it is safe for the file provider to remove the content file.
         // Care should be taken that the corresponding placeholder file stays behind after the content file has been deleted.
-        
+
         // Called after the last claim to the file has been released. At this point, it is safe for the file provider to remove the content file.
-        
+
         // TODO: look up whether the file has local changes
         let fileHasLocalChanges = false
-        
+
         if !fileHasLocalChanges {
             // remove the existing file to free up space
             do {
@@ -141,30 +138,30 @@ class FileProviderExtension: NSFileProviderExtension {
             } catch {
                 // Handle error
             }
-            
+
             // write out a placeholder to facilitate future property lookups
-            self.providePlaceholder(at: url, completionHandler: { error in
+            self.providePlaceholder(at: url, completionHandler: { _ in
                 // TODO: handle any error, do any necessary cleanup
             })
         }
     }
-    
+
     // MARK: - Actions
-    
+
     /* TODO: implement the actions for items here
      each of the actions follows the same pattern:
      - make a note of the change in the local model
      - schedule a server request as a background task to inform the server of the change
      - call the completion block with the modified item in its post-modification state
      */
-    
+
     // MARK: - Enumeration
-    
+
     override func enumerator(for containerItemIdentifier: NSFileProviderItemIdentifier) throws -> NSFileProviderEnumerator {
         let maybeEnumerator: NSFileProviderEnumerator? = nil
-        if (containerItemIdentifier == NSFileProviderItemIdentifier.rootContainer) {
+        if containerItemIdentifier == NSFileProviderItemIdentifier.rootContainer {
             // TODO: instantiate an enumerator for the container root
-        } else if (containerItemIdentifier == NSFileProviderItemIdentifier.workingSet) {
+        } else if containerItemIdentifier == NSFileProviderItemIdentifier.workingSet {
             // TODO: instantiate an enumerator for the working set
         } else {
             // TODO: determine if the item is a directory or a file
@@ -172,9 +169,8 @@ class FileProviderExtension: NSFileProviderExtension {
             // - for a file, instantiate an enumerator that observes changes to the file
         }
         guard let enumerator = maybeEnumerator else {
-            throw NSError(domain: NSCocoaErrorDomain, code: NSFeatureUnsupportedError, userInfo:[:])
+            throw NSError(domain: NSCocoaErrorDomain, code: NSFeatureUnsupportedError, userInfo: [:])
         }
         return enumerator
     }
-    
 }
